Add optional image to connectors and cables (#153)
This image, with an optional caption below, is displayed in the lower
section of the connector/cable node in the diagram - just above the
notes if present.

This solves the basic part of issue #27, and is a continuation of
PR #137 that was closed due to changes in the base branch.